+++ src/api/calendarEvents.ts
@@ -1,17 +1,54 @@
+import axios from "axios"
 import type { GatsbyFunctionRequest, GatsbyFunctionResponse } from "gatsby"
-import { lambda } from "../lambda/calendarEvents"
 
 async function handler(
   __req: GatsbyFunctionRequest,
   res: GatsbyFunctionResponse
 ): Promise<void> {
-  // passing env vars as arguments due to a bug on GC functions where env vars
-  // can not be accessed by imported functions
-  const { statusCode, body } = await lambda(
-    process.env.GOOGLE_API_KEY!,
-    process.env.GOOGLE_CALENDAR_ID!
-  )
-  res.status(statusCode).send(body)
+  const apiKey = process.env.GOOGLE_API_KEY
+  const calendarId = process.env.GOOGLE_CALENDAR_ID
+
+  try {
+    const futureEventsReq = await axios.get(
+      `https://content.googleapis.com/calendar/v3/calendars/${calendarId}/events`,
+      {
+        params: {
+          key: apiKey,
+          timeMin: new Date().toISOString(),
+          maxResults: 3,
+          singleEvents: true,
+          orderBy: "startTime",
+        },
+      }
+    )
+
+    const pastEventsReq = await axios.get(
+      `https://content.googleapis.com/calendar/v3/calendars/${calendarId}/events`,
+      {
+        params: {
+          key: apiKey,
+          timeMax: new Date().toISOString(),
+          maxResults: 4,
+          singleEvents: true,
+          orderBy: "startTime",
+        },
+      }
+    )
+
+    const response = {
+      pastEvents: pastEventsReq.data.items,
+      futureEvents: futureEventsReq.data.items,
+    }
+
+    res.status(200).send(JSON.stringify(response))
+  } catch (error) {
+    console.error(error)
+    res.status(500).send(
+      JSON.stringify({
+        msg: "Something went wrong with requesting the calendar events data.",
+      })
+    )
+  }
 }
 
 export default handler
